requirements.txt
python-telegram-bot==20.7
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0
numpy==1.26.2
matplotlib==3.8.2
scikit-learn==1.3.2
reportlab==4.0.8
validators==0.22.0 